requirements.txt
aiokafka==0.12.0
asyncpg==0.30.0
fastapi==0.128.0
kafka_python==2.2.4
Pillow==12.1.0
psycopg2_binary==2.9.9
pydantic==2.12.5
python-dotenv==1.2.1
python_jose==3.4.0
uvicorn==0.40.0
python-multipart==0.0.20
